backend/requirements.txt
python-dotenv
langgraph
langchain
langchain-openai
langchain_google_genai
langgraph-cli